% Generated by roxygen2 (4.0.2): do not edit by hand
\name{allanwood}
\alias{allanwood}
\title{Query http://www.alanwood.net/pesticides}
\usage{
allanwood(x, verbose = TRUE)
}
\arguments{
\item{x}{character; A common name.}

\item{verbose}{logical; print message during processing to console?}
}
\value{
character; The CAS number. NA if the substance is not found.
}
\description{
Query http://www.alanwood.net/pesticides for CAS and activity
}
\note{
If more than one reference is found on the first is taken.
}
\examples{
allanwood('Fluazinam')
sapply(c('Fluazinam', 'Diclofop', 'xxxxx'), allanwood)
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

