% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cactus}
\alias{cactus}
\title{Query cactus}
\usage{
cactus(query, output = "smiles", verbose = FALSE, ...)
}
\arguments{
\item{query}{character; as vector of search queries (e.g. CAS numbers).}

\item{output}{character; what output should be returned ? Available options are 'smiles',
'iupac_name', 'mw' (molecular weight), see \url{http://cactus.nci.nih.gov/chemical/structure_documentation} for more options.}

\item{verbose}{logical; should a verbose output be printed on the console?}

\item{...}{currently not used.}
}
\value{
A character vector.
}
\description{
Query cactus (\url{http://cactus.nci.nih.gov/chemical/structure_documentation}).
}
\note{
If more than one hit is found, only the first is returned. A warinig is given.
}
\examples{
# convert CAS to smiles
casnr <- c("107-06-2", "107-13-1", "319-84-6", "319-86-8", "1031-07-8")
cactus(casnr, output = 'smiles')
# Get molecular weight from CAS
casnr <- c("107-06-2", "107-13-1", "319-84-6")
cactus(casnr, output = 'mw', verbose = TRUE)
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

