% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pubchem.R
\name{cid_to_ext}
\alias{cid_to_ext}
\title{Get extended information from Pubchem}
\usage{
cid_to_ext(cid, verbose = FALSE, ...)
}
\arguments{
\item{cid}{character, CID as returned by get_cid.}

\item{verbose}{logical; should a verbose output be printed on the console?}

\item{...}{currently not used.}
}
\value{
a charater vector
}
\description{
Get extended info from pubchem, see \url{https://pubchem.ncbi.nlm.nih.gov/}
}
\examples{
# convert CAS to CID
casnr <- c("107-06-2", "107-13-1", "319-84-6", "319-86-8")
cid <- get_cid(casnr)
# get SMILES from CID
cid_to_ext(cid)
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

