% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cid_to_smiles}
\alias{cid_to_smiles}
\title{Convert CID to SMILES}
\usage{
cid_to_smiles(cid, verbose = FALSE, ...)
}
\arguments{
\item{cid}{character, CID as returned by get_cid.}

\item{verbose}{logical; should a verbose output be printed on the console?}

\item{...}{currently not used.}
}
\value{
a charater vector.
}
\description{
Convert CompoundID (CID) to SMILES, see \url{https://pubchem.ncbi.nlm.nih.gov/}
}
\examples{
# convert CAS to CID
casnr <- c("107-06-2", "107-13-1", "319-84-6", "319-86-8")
cid <- get_cid(casnr)
# get SMILES from CSID
cid_to_smiles(cid)
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

