% Generated by roxygen2 (4.0.1): do not edit by hand
\name{csid_to_smiles}
\alias{csid_to_smiles}
\title{Convert CSID to SMILES via chemspider}
\usage{
csid_to_smiles(csid, token, verbose = FALSE, ...)
}
\arguments{
\item{csid}{character, CSID as returned by get_csid.}

\item{token}{character; security token.}

\item{verbose}{logical; should a verbose output be printed on the console?}

\item{...}{currently not used.}
}
\value{
a charater vector of class 'csid' with CSID.
}
\description{
Convert ChemspiderId (CSID) to SMILES, see \url{http://www.chemspider.com/}
}
\note{
A security token is neeeded. Please register at RSC
\url{https://www.rsc.org/rsc-id/register}
for a security token.
}
\examples{
token <- '37bf5e57-9091-42f5-9274-650a64398aaf'
# convert CAS to CSID
casnr <- c("107-06-2", "107-13-1", "319-84-6", "319-86-8", "1031-07-8")
csid <- get_csid(casnr, token = token)
# get SMILES from CSID
csid_to_smiles(csid, token)
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

