% Generated by roxygen2 (4.0.2): do not edit by hand
\name{etox_to_cas}
\alias{etox_to_cas}
\title{Search for CAS numbers on ETOX}
\usage{
etox_to_cas(x)
}
\arguments{
\item{x}{character; The searchterm}
}
\value{
character; The CAS number. NA if the substance is not found.
#'
}
\description{
Search for CAS numbers on ETOX (https://webetox.uba.de/webETOX/index.do)
}
\examples{
etox_to_cas('2,4,5-Trichlorphenol')
etox_to_cas('xxxxx')
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

