% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.R
\name{geomean}
\alias{geomean}
\title{Geometric mean}
\usage{
geomean(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector.}

\item{na.rm}{a logical value indicating whether NA values should be 
stripped before the computation proceeds.}
}
\value{
numeric vector of length one with the geometric mean.
}
\description{
Function for the geometric mean.
}
\details{
The geometric mean is comuted as
\deqn{x = e^{(\sum \log x) / n}}
}
\examples{
x <- c(1, 10, 100)
mean(x)
geomean(x)
}
\references{
\url{http://stackoverflow.com/questions/2602583/geometric-mean-is-there-a-built-in}
}
