% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_csid}
\alias{get_csid}
\title{Get CSID from chemspider}
\usage{
get_csid(query, token, verbose = FALSE, ...)
}
\arguments{
\item{query}{charachter; search query (e.g. CAS numbers).}

\item{token}{character; security token.}

\item{verbose}{logical; should a verbose output be printed on the console?}

\item{...}{currently not used.}
}
\value{
a character vector of class 'csid' with CSID.
}
\description{
Return ChemspiderId (CSID) for a search query, see \url{http://www.chemspider.com/}.
}
\note{
A security token is neeeded. Please register at RSC
\url{https://www.rsc.org/rsc-id/register}
for a security token.
}
\examples{
token <- '37bf5e57-9091-42f5-9274-650a64398aaf'
casnr <- c("107-06-2", "107-13-1", "319-84-6", "319-86-8", "1031-07-8")
get_csid(casnr, token = token)
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

