% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_kow}
\alias{get_kow}
\title{Retrieve KOW from CAS}
\usage{
get_kow(cas, verbose = FALSE, ...)
}
\arguments{
\item{cas}{charachter; CAS numbers.}

\item{verbose}{logical; should a verbose output be printed on the console?}

\item{...}{currently not used.}
}
\value{
a character vector.
}
\description{
Retrieve KOW from CAS via LOGKOW, see \url{http://logkow.cisti.nrc.ca/logkow/search.html}.
}
\note{
If more then on match is found user is asked for input.
}
\examples{
casnr <- c("107-06-2", "107-13-1", "319-84-6", "319-86-8", "1918-16-7")
get_kow(casnr, verbose = TRUE)
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

