% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_edi.R
\name{theme_edi}
\alias{theme_edi}
\title{Custom ggplot2 theme}
\usage{
theme_edi(base_size = 14, base_family = "Helvetica")
}
\arguments{
\item{base_size}{basefont size}

\item{base_family}{base font family}
}
\description{
Custom ggplot2 theme
}
\examples{
library(ggplot2)
p <- ggplot(mtcars) + 
 geom_point(aes(x = wt, y = mpg, 
colour=factor(gear))) + facet_wrap(~am)
p
p + theme_edi()
}

