% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_number.R
\name{to_number}
\alias{to_number}
\title{translate English spelled quantities into their integer counterparts.}
\usage{
to_number(x)
}
\arguments{
\item{x}{A English word for a number.}
}
\description{
Allows you to translate to integer numerical words spelled in English
Text must be previously cleaned & removed extraneous words or symbols
Quantities MUST be written in a correct English (this is not a grammar tool)
The upper limit is up to the millions range. Cents must be removed.
}
\examples{
to_number("one thousand and seventy two")
to_number(c("thirty seven", "forty two"))

}
\keyword{currency,}
\keyword{euros}
\keyword{money,}
