% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words2number.R
\docType{package}
\name{words2number}
\alias{words2number}
\alias{words2number-package}
\title{words2number: A package to convert number words to numeric digits}
\description{
This package provides one important function:
to_number()
}
\section{to_number()}{

Translate English spelled quantities into their integer counterparts.
Allows you to translate to integer numerical words spelled in English.
Text must be previously cleaned & removed extraneous words or symbols.
Quantities MUST be written in a correct English. (this is not a grammar tool)
The upper limit is up to the millions range. Cents must be removed.
}

