% Generated by roxygen2 (4.0.2): do not edit by hand
\name{etox_to_cas}
\alias{etox_to_cas}
\title{Search for CAS numbers on ETOX}
\usage{
etox_to_cas(x, verbose = TRUE)
}
\arguments{
\item{x}{character; The searchterm}

\item{verbose}{logical; print message during processing to console?}
}
\value{
character; The CAS number. NA if the substance is not found.
}
\description{
Search for CAS numbers on ETOX (https://webetox.uba.de/webETOX/index.do)
}
\note{
If more than one reference is found on the first is taken.
}
\examples{
etox_to_cas('2,4,5-Trichlorphenol')
sapply(c('2,4,5-Trichlorphenol', '2,4-D', 'xxxxx'), etox_to_cas)
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

