% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_regnie.R
\name{read_regnie}
\alias{read_regnie}
\title{Read DWD REGNIE gridded data into R}
\usage{
read_regnie(file)
}
\arguments{
\item{file}{path to gz archive}
}
\value{
A \code{RasterLayer} object.
}
\description{
This functions reads DWD gridded data (currently only tested with REGNIE).
A description of the data can be found here:
\url{ftp://ftp.dwd.de/pub/CDC/grids_germany/daily/regnie/REGNIE_Beschreibung.pdf}.
Data is available here: \url{ftp://ftp.dwd.de/pub/CDC/grids_germany/daily/regnie/}.
}
\examples{
# Read daily precipitation on 24.12.2014.
r <- read_regnie(system.file("extdata", "ra141224.gz", package = "esmisc"))
}

